SVM: Loader: absorb rest of `program_loader`
