+++ svm/src/lib.rs
@@ -6,7 +6,6 @@ pub mod account_rent_state;
 pub mod loader;
 pub mod message_processor;
 pub mod nonce_info;
-pub mod program_loader;
 pub mod runtime_config;
 pub mod transaction_account_state_info;
 pub mod transaction_error_metrics;
